+++ r/R/thiessen-package.R
@@ -1,3 +1,11 @@
 #' @keywords internal
+#' @references
+#' Chipman, H. A., George, E. I. and McCulloch, R. E. (2010). BART: Bayesian
+#' additive regression trees. *The Annals of Applied Statistics* 4(1),
+#' 266-298. \doi{10.1214/09-AOAS285}
+#'
+#' Stone, A. and Gosling, J. P. (2025). AddiVortes: (Bayesian) additive
+#' Voronoi tessellations. *Journal of Computational and Graphical
+#' Statistics* 34(3), 859-871. \doi{10.1080/10618600.2024.2414104}
 #' @useDynLib thiessen, .registration = TRUE
 "_PACKAGE"